Module aws_sdk_amplifybackend::error

Modules

bad_request_exception

See BadRequestException

gateway_timeout_exception

See GatewayTimeoutException

not_found_exception

See NotFoundException

too_many_requests_exception

See TooManyRequestsException

Structs

BadRequestException

An error returned if a request is not formed properly.

CloneBackendError

CreateBackendAPIError

CreateBackendAuthError

CreateBackendConfigError

CreateBackendError

CreateTokenError

DeleteBackendAPIError

DeleteBackendAuthError

DeleteBackendError

DeleteTokenError

GatewayTimeoutException

An error returned if there's a temporary issue with the service.

GenerateBackendAPIModelsError

GetBackendAPIError

GetBackendAPIModelsError

GetBackendAuthError

GetBackendError

GetBackendJobError

GetTokenError

ImportBackendAuthError

ListBackendJobsError

NotFoundException

An error returned when a specific resource type is not found.

RemoveAllBackendsError

RemoveBackendConfigError

TooManyRequestsException

An error that is returned when a limit of a specific type has been exceeded.

UpdateBackendAPIError

UpdateBackendAuthError

UpdateBackendConfigError

UpdateBackendJobError

Enums

CloneBackendErrorKind

CreateBackendAPIErrorKind

CreateBackendAuthErrorKind

CreateBackendConfigErrorKind

CreateBackendErrorKind

CreateTokenErrorKind

DeleteBackendAPIErrorKind

DeleteBackendAuthErrorKind

DeleteBackendErrorKind

DeleteTokenErrorKind

GenerateBackendAPIModelsErrorKind

GetBackendAPIErrorKind

GetBackendAPIModelsErrorKind

GetBackendAuthErrorKind

GetBackendErrorKind

GetBackendJobErrorKind

GetTokenErrorKind

ImportBackendAuthErrorKind

ListBackendJobsErrorKind

RemoveAllBackendsErrorKind

RemoveBackendConfigErrorKind

UpdateBackendAPIErrorKind

UpdateBackendAuthErrorKind

UpdateBackendConfigErrorKind

UpdateBackendJobErrorKind