Module aws_sdk_amplifybackend::model

Modules

backend_api_app_sync_auth_settings

See BackendApiAppSyncAuthSettings

backend_api_auth_type

See BackendApiAuthType

backend_api_conflict_resolution

See BackendApiConflictResolution

backend_api_resource_config

See BackendApiResourceConfig

backend_auth_apple_provider_config

See BackendAuthAppleProviderConfig

backend_auth_social_provider_config

See BackendAuthSocialProviderConfig

backend_job_resp_obj

See BackendJobRespObj

create_backend_auth_forgot_password_config

See CreateBackendAuthForgotPasswordConfig

create_backend_auth_identity_pool_config

See CreateBackendAuthIdentityPoolConfig

create_backend_auth_mfa_config

See CreateBackendAuthMfaConfig

create_backend_auth_o_auth_config

See CreateBackendAuthOAuthConfig

create_backend_auth_password_policy_config

See CreateBackendAuthPasswordPolicyConfig

create_backend_auth_resource_config

See CreateBackendAuthResourceConfig

create_backend_auth_user_pool_config

See CreateBackendAuthUserPoolConfig

email_settings

See EmailSettings

login_auth_config_req_obj

See LoginAuthConfigReqObj

resource_config

See ResourceConfig

settings

See Settings

sms_settings

See SmsSettings

social_provider_settings

See SocialProviderSettings

update_backend_auth_forgot_password_config

See UpdateBackendAuthForgotPasswordConfig

update_backend_auth_identity_pool_config

See UpdateBackendAuthIdentityPoolConfig

update_backend_auth_mfa_config

See UpdateBackendAuthMfaConfig

update_backend_auth_o_auth_config

See UpdateBackendAuthOAuthConfig

update_backend_auth_password_policy_config

See UpdateBackendAuthPasswordPolicyConfig

update_backend_auth_resource_config

See UpdateBackendAuthResourceConfig

update_backend_auth_user_pool_config

See UpdateBackendAuthUserPoolConfig

Structs

BackendApiAppSyncAuthSettings

The authentication settings for accessing provisioned data models in your Amplify project.

BackendApiAuthType

Describes the auth types for your configured data models.

BackendApiConflictResolution

Describes the conflict resolution configuration for your data model configured in your Amplify project.

BackendApiResourceConfig

The resource config for the data model, configured as a part of the Amplify project.

BackendAuthAppleProviderConfig

Describes Apple social federation configurations for allowing your app users to sign in using OAuth.

BackendAuthSocialProviderConfig

Describes third-party social federation configurations for allowing your app users to sign in using OAuth.

BackendJobRespObj

The response object for this operation.

CreateBackendAuthForgotPasswordConfig

Describes the forgot password policy for authenticating into the Amplify app.

CreateBackendAuthIdentityPoolConfig

Describes authorization configurations for the auth resources, configured as a part of your Amplify project.

CreateBackendAuthMfaConfig

Describes whether to apply multi-factor authentication policies for your Amazon Cognito user pool configured as a part of your Amplify project.

CreateBackendAuthOAuthConfig

Creates the OAuth configuration for your Amplify project.

CreateBackendAuthPasswordPolicyConfig

The password policy configuration for the backend to your Amplify project.

CreateBackendAuthResourceConfig

Defines the resource configuration when creating an auth resource in your Amplify project.

CreateBackendAuthUserPoolConfig

Describes the Amazon Cognito user pool configuration for the auth resource to be configured for your Amplify project.

EmailSettings

The configuration for the email sent when an app user forgets their password.

LoginAuthConfigReqObj

The request object for this operation.

ResourceConfig

Defines the resource configuration for the data model in your Amplify project.

Settings

The settings of your MFA configuration for the backend of your Amplify project.

SmsSettings

SMS settings for authentication.

SocialProviderSettings

The settings for using the social identity providers for access to your Amplify app.

UpdateBackendAuthForgotPasswordConfig

Describes the forgot password policy for authenticating into the Amplify app.

UpdateBackendAuthIdentityPoolConfig

Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

UpdateBackendAuthMfaConfig

Updates the multi-factor authentication (MFA) configuration for the backend of your Amplify project.

UpdateBackendAuthOAuthConfig

The OAuth configurations for authenticating users into your Amplify app.

UpdateBackendAuthPasswordPolicyConfig

Describes the password policy for your Amazon Cognito user pool configured as a part of your Amplify project.

UpdateBackendAuthResourceConfig

Defines the resource configuration when updating an authentication resource in your Amplify project.

UpdateBackendAuthUserPoolConfig

Describes the Amazon Cognito user pool configuration for the authorization resource to be configured for your Amplify project on an update.

Enums

AdditionalConstraintsElement

AuthResources

DeliveryMethod

MfaMode

MfaTypesElement

Mode

OAuthGrantType

OAuthScopesElement

RequiredSignUpAttributesElement

ResolutionStrategy

Service

SignInMethod

Status