Module aws_sdk_amplifybackend::operation

Structs

CloneBackend

This operation clones an existing backend.

CreateBackend

This operation creates a backend for an Amplify app. Backends are automatically created at the time of app creation.

CreateBackendAPI

Creates a new backend API resource.

CreateBackendAuth

Creates a new backend authentication resource.

CreateBackendConfig

Creates a config object for a backend.

CreateToken

Generates a one-time challenge code to authenticate a user into your Amplify Admin UI.

DeleteBackend

Removes an existing environment from your Amplify project.

DeleteBackendAPI

Deletes an existing backend API resource.

DeleteBackendAuth

Deletes an existing backend authentication resource.

DeleteToken

Deletes the challenge token based on the given appId and sessionId.

GenerateBackendAPIModels

Generates a model schema for an existing backend API resource.

GetBackend

Provides project-level details for your Amplify UI project.

GetBackendAPI

Gets the details for a backend API.

GetBackendAPIModels

Generates a model schema for existing backend API resource.

GetBackendAuth

Gets a backend auth details.

GetBackendJob

Returns information about a specific job.

GetToken

Gets the challenge token based on the given appId and sessionId.

ImportBackendAuth

Imports an existing backend authentication resource.

ListBackendJobs

Lists the jobs for the backend of an Amplify app.

RemoveAllBackends

Removes all backend environments from your Amplify project.

RemoveBackendConfig

Removes the AWS resources required to access the Amplify Admin UI.

UpdateBackendAPI

Updates an existing backend API resource.

UpdateBackendAuth

Updates an existing backend authentication resource.

UpdateBackendConfig

Updates the AWS resources required to access the Amplify Admin UI.

UpdateBackendJob

Updates a specific job.